requirements.txt
pymongo==4.3.3
pymongo[srv]
python-dotenv